refactor: move the image-pair rule into the env schema
Follow-up to the previous commit. The pair gate was a
`if (getExecutionMode() === "docker")` in bootCritical — a backend-id
branch, which is exactly the closed list of backends the orchestrator
registry exists to keep open (it replaced a compile-time-closed
`Record<ExecutionMode, …>` for that reason).

The rule is a property of two env values, not of a backend, so it now
lives where every other cross-field env rule lives: a `.refine()` on the
env schema, next to the S3_BUCKET/S3_REGION and APP_URL ones. No
condition on RUN_ADAPTER at all — the docker orchestrator reads both
refs at run time, and the firecracker rootfs build reads both from the
same environment to bake a guest image, so a mismatched pair is a bug
for both; backends that ignore the vars inherit matching defaults, so
the rule is a no-op for them.

`parseImageRef` / `findImageTagMismatch` move to a new
`@appstrate/core/image-ref` (env already depends on core; core does not
depend on env). `services/orchestrator/runtime-image-pair.ts` keeps only
the half that needs a live Docker host — the revision-drift warning.

Simplifications in the same pass:

- release.yml build-args reverted. Verified against the published
  1.0.0-beta.51 sidecar: metadata-action's `labels:` input already
  stamps a real `org.opencontainers.image.revision` (a `--label`
  overrides the Dockerfile `LABEL`), so the args were dead weight. The
  comment now records why, and what would break if `labels:` went away.
  preview.yml has no such input and does need them.
- The build script no longer feeds BUILD_VERSION. Nothing reads
  `org.opencontainers.image.version`, and a local build has no release
  version — that dropped a `git describe --match "v*"` and its rationale.
- Trimmed the doc duplicated between the core module and the api one.
