+++ .github/workflows/preview.yml
@@ -130,14 +130,12 @@ jobs:
           # staging/prod server — skipping arm64 roughly halves build time.
           platforms: linux/amd64
           push: true
-          # Stamp `org.opencontainers.image.{version,revision,created}` into
-          # runtime-pi + sidecar. The platform compares the two `revision`
-          # labels after pre-pulling and warns when the pair was built from
-          # different commits (#1200); an unstamped preview image reads
-          # `unknown` on both sides and silently disables that check — which is
-          # worst here, since preview deploys are exactly where someone points
-          # the two env vars at two different tags. Images that declare none of
-          # these ARGs ignore them (docker warns, never fails).
+          # Unlike release.yml, this job has no metadata-action `labels:` input,
+          # so runtime-pi + sidecar get their OCI labels from these ARGs. The
+          # platform compares the two `org.opencontainers.image.revision` labels
+          # and warns on a pair built from different commits (#1200); unstamped
+          # preview images read `unknown` on both sides and disable that check
+          # silently. Images declaring no such ARG ignore them.
           build-args: |
             BUILD_VERSION=${{ needs.resolve.outputs.tag }}
             BUILD_REVISION=${{ github.sha }}
