+++ .github/workflows/release.yml
@@ -163,22 +163,19 @@ jobs:
           file: ${{ matrix.dockerfile }}
           platforms: linux/amd64,linux/arm64
           push: true
-          # Stamp the deployed build identity into the image. Two naming
-          # conventions, because two audiences:
-          #   APP_VERSION / GIT_SHA — the main Dockerfile only, surfaced as env
-          #     vars via /health and the SPA footer.
-          #   BUILD_* — runtime-pi + sidecar, which turn them into
-          #     `org.opencontainers.image.{version,revision,created}` labels.
-          #     The platform compares the two `revision` labels after
-          #     pre-pulling and warns on a mismatched pair (#1200), so an
-          #     unstamped release image would silently disable that check.
-          # Images declaring neither set ignore both (docker warns, never fails).
+          # Stamp the deployed build identity into the image. Only the main
+          # Dockerfile declares these ARGs; other matrix images ignore them.
+          #
+          # The runtime-pi / sidecar images need no build-arg of their own: the
+          # `labels:` input below carries metadata-action's OCI set, and a
+          # `--label` overrides the Dockerfile's `LABEL`, so released images
+          # already carry a real `org.opencontainers.image.revision` (verified
+          # on 1.0.0-beta.51). That label is what the platform compares to warn
+          # on a mismatched runtime/sidecar pair (#1200) — do not drop the
+          # `labels:` input without feeding BUILD_REVISION instead.
           build-args: |
             APP_VERSION=${{ github.ref_name }}
             GIT_SHA=${{ github.sha }}
-            BUILD_VERSION=${{ github.ref_name }}
-            BUILD_REVISION=${{ github.sha }}
-            BUILD_CREATED=${{ github.event.repository.updated_at }}
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
           cache-from: type=gha
