+++ CLAUDE.md
@@ -92,7 +92,7 @@ User Browser (BrowserRouter SPA)  Platform (Bun + Hono :3000)
 
 Sidecar + agent setup run in parallel (`Promise.all`). Images pre-pulled at boot (`ensureImage`) and kept warm afterwards by the runtime-image warmer (`services/orchestrator/runtime-image-warmer.ts`: reconciles one `appstrate-imagepin-*` holder container per image, so host-level `docker image prune -a` can't put a cold pull back on the run-boot path). Full sidecar protocol: `docs/architecture/SIDECAR.md`.
 
-`PI_IMAGE` and `SIDECAR_IMAGE` are a **version contract**, not two independent knobs: the agent runtime and the sidecar speak a wire protocol that changes in the same commit, and a mismatched pair boots fine then fails runs with an opaque upstream error (#1195). Two guards, both in `services/orchestrator/runtime-image-pair.ts`: the boot **fails** when the two refs are pinned to different tags, and after the pre-pull the platform **warns** when the two images on the host carry different `org.opencontainers.image.revision` stamps. `bun run docker:build:runtime` (alias `build-runtime`) builds BOTH images with the same stamp — there is deliberately no command that rebuilds one half.
+`PI_IMAGE` and `SIDECAR_IMAGE` are a **version contract**, not two independent knobs: the agent runtime and the sidecar speak a wire protocol that changes in the same commit, and a mismatched pair boots fine then fails runs with an opaque upstream error (#1195). Two guards. The env schema (`@appstrate/env`, via `findImageTagMismatch` from `@appstrate/core/image-ref`) **fails boot** when the two refs carry different tags — a cross-field rule like the S3/APP_URL ones, not conditioned on `RUN_ADAPTER`. After the pre-pull, `services/orchestrator/runtime-image-pair.ts` **warns** when the two images on the host carry different `org.opencontainers.image.revision` stamps (same tag, two builds — `:latest` rebuilt on one side only). `bun run docker:build:runtime` (alias `build-runtime`) builds BOTH images with the same stamp — there is deliberately no command that rebuilds one half.
 
 **Run liveness is two-phase** (`services/run-watchdog.ts`, same split as Kubernetes `startupProbe` vs `livenessProbe`): until the runner's first event the platform is provisioning and attests liveness on its behalf (`services/run-boot-heartbeat.ts`), bounded by `runs.boot_deadline_at` (`RUN_BOOT_DEADLINE_SECONDS`, default 300s); after it, the runner's own heartbeat owns liveness (`RUN_STALL_THRESHOLD_SECONDS`, default 60s). Each predicate finalises with its own error.
