+++ packages/core/package.json
@@ -56,6 +56,7 @@
     "./zip": "./src/zip.ts",
     "./package-files": "./src/package-files.ts",
     "./naming": "./src/naming.ts",
+    "./image-ref": "./src/image-ref.ts",
     "./mime": "./src/mime.ts",
     "./dependencies": "./src/dependencies.ts",
     "./integrity": "./src/integrity.ts",
